Code, Files, Documentation
Modified code for Scorpio, making the Right-Scorpio code's autonomous more efficient and precise. Edited the documentation of all three robots, adding images and describing what the robots are capable of in a short paragraph. Added AppendixA of the Tipping Point rules, which contains field specifications, which can be used in both planning and coding.